+++ src/main/java/com/meli/desafio_spring/users/models/Buyer.java
@@ -4,22 +4,22 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Buyer extends User {
-    private List<Long> followed;
+    private List<Integer> followed;
 
-    public Buyer(long userId, String userName) {
+    public Buyer(int userId, String userName) {
         super(userId, userName);
         this.followed = new ArrayList<>();
     }
 
-    public List<Long> getFollowed() {
+    public List<Integer> getFollowed() {
         return followed;
     }
 
-    public void setFollowed(List<Long> followed) {
+    public void setFollowed(List<Integer> followed) {
         this.followed = followed;
     }
 
-    public void follow(Long sellerId) {
+    public void follow(Integer sellerId) {
         followed.add(sellerId);
     }
 }
