+++ src/main/java/com/meli/desafio_spring/users/models/Seller.java
@@ -4,22 +4,22 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class Seller extends User {
-    private List<Long> followers;
+    private List<Integer> followers;
 
-    public Seller(long userId, String userName) {
+    public Seller(int userId, String userName) {
         super(userId, userName);
         this.followers = new ArrayList<>();
     }
 
-    public List<Long> getFollowers() {
+    public List<Integer> getFollowers() {
         return followers;
     }
 
-    public void setFollowers(List<Long> followers) {
+    public void setFollowers(List<Integer> followers) {
         this.followers = followers;
     }
 
-    public void addFollower(Long buyerId) {
+    public void addFollower(Integer buyerId) {
         this.followers.add(buyerId);
     }
 }
